+++ src/main/java/com/cout970/magneticraft/items/ItemToolCharger.java
@@ -48,7 +48,7 @@ public class ItemToolCharger extends ItemCharged {
                         }
                     } else if (Magneticraft.COFH_ENERGY && (it instanceof IEnergyContainerItem)) {//calcs in RF
                         IEnergyContainerItem st = (IEnergyContainerItem) it;
-                        int space = (int) (st.getMaxEnergyStored(s) - st.getEnergyStored(s));
+                        int space = st.getMaxEnergyStored(s) - st.getEnergyStored(s);
                         int toMove = (int) Math.min(space, EnergyConverter.WtoRF(getCharge(item)));
                         if (toMove > 0) {
                             st.receiveEnergy(s, toMove, false);
